Avoid warnings of pedantic compiler options
git-svn-id: svn://svn.code.sf.net/p/qwt/code/branches/qwt-6.0@1011 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ src/qwt_curve_fitter.cpp
@@ -131,7 +131,7 @@ int QwtSplineCurveFitter::splineSize() const
 */
 QPolygonF QwtSplineCurveFitter::fitCurve( const QPolygonF &points ) const
 {
-    const int size = ( int )points.size();
+    const int size = points.size();
     if ( size <= 2 )
         return points;
